+++ .gitignore
@@ -1,4 +1,4 @@
-src/__pycache__
+ergosafe/__pycache__
 __pycache__
 cam1.mp4
 cam2.mp4
