+++ Dockerfile
@@ -46,7 +46,7 @@ COPY requirements.txt .
 RUN pip install --no-cache-dir -r requirements.txt
 
 # Copiar código-fonte
-COPY ./src ./src
+COPY ./ergosafe ./ergosafe
 
 # Comando por defeito (pode ser alterado via docker-compose)
-CMD ["uvicorn", "src.main:app", "--host", "0.0.0.0", "--port", "8000", "--reload"]
+CMD ["uvicorn", "ergosafe.main:app", "--host", "0.0.0.0", "--port", "8000", "--reload"]
